Extracted Searchbar into component
