+++ src/components/index.ts
@@ -32,6 +32,7 @@ import SecondaryNavBar from './Nav/SecondaryNavBar'
 // Selects
 import Select from './Selects/Select'
 import SearchBar from './Selects/SearchBar'
+import InviteAdminSearchBar from './Selects/InviteAdminSearchBar'
 // Text
 import NumberCircle from './Text/NumberCircle'
 import HighlightedTitle from './Text/HighlightedTitle'
@@ -84,4 +85,5 @@ export {
   ClimateClock,
   AdminBox,
   AdminListTable,
+  InviteAdminSearchBar,
 }
